backend/requirements.txt
asgiref>=3.8.1
cffi
cryptography
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
dnspython
drf-yasg
inflection
packaging
pycparser
PyJWT
pymongo
PyMySQL
python-dotenv
pytz
PyYAML
sqlparse
typing_extensions
uritemplate
requests
